arreglado el problema de ids int a Long

+++ src/main/java/com/contactos/rest/contactos/repository/ContactoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.contactos.rest.contactos.model.Contacto;
 
-public interface ContactoRepository extends JpaRepository<Contacto, Integer>{
+public interface ContactoRepository extends JpaRepository<Contacto, Long>{
 
 }
